move setVersion to reusable module
